embodied_active_learning_core/requirements.txt
numpy
opencv-python
Pillow
scikit-image>=0.14.0
torch
git+https://github.com/drsleep/DenseTorch.git
albumentations
setuptools
tensorflow
tf
rospy
sklearn
scikit-learn
PyYAML
open3d
scipy
airsim
pandas
torchvision
catkin